Add files via upload
I created this project to display an up-to-date bar chart of the most popular projects on Github according to Stars. I began by using the requests package in python to get some information in real-time. The function uses a GET request to ping Githubs API and returns a .json of the most-starred repositories. I then passed the results into a python dictionary. Then I created a loop function to parse through the dictionary and return the information I wanted to present in the bar chart. After the loop function created 2 lists stored as variables, I passed those variables to streamlit to create a bar chart using plotly express and display the information on the local machine's web browser that can be accessed using port 8501. Every time the webpage is visited it executes the function and returns an up-to-date chart with repository name and star count.